fix(payments): cast SUM(amount_minor) to bigint before decoding
Postgres's SUM(bigint) returns numeric, not bigint, to avoid silent
overflow. Two refund-summing queries decoded that column straight into
i64, which sqlx rejects at runtime with a type mismatch — confirmed live:
every create_refund call failed with a generic internal_error. Both now
cast the aggregate back to ::bigint to match the declared amount_minor
columns, which are always within i64 range in practice.

+++ crates/chaos-core/src/adapters/postgres/payments/events.rs
@@ -131,7 +131,7 @@ async fn recompute_order_refund_summary(
         "SELECT COALESCE((SELECT SUM(refund.amount_minor) FROM commerce.refunds AS refund \
                            WHERE refund.store_id = sales_order.store_id \
                              AND refund.order_id = sales_order.id \
-                             AND refund.status = 'succeeded'), 0), \
+                             AND refund.status = 'succeeded'), 0)::bigint, \
                 sales_order.total_amount_minor \
          FROM commerce.orders AS sales_order \
          WHERE sales_order.store_id = $1 AND sales_order.id = $2",
